+++ castle/cms/tiles/gallery.py
@@ -1,11 +1,10 @@
 from castle.cms.tiles.base import BaseImagesTile
 from castle.cms.tiles.base import DisplayTypeTileMixin
+from castle.cms.tiles.content import IImagesTileSchema
 from castle.cms.tiles.views import BaseTileView
 from castle.cms.tiles.views import TileViewsSource
-from castle.cms.widgets import ImageRelatedItemsFieldWidget
 from castle.cms.widgets import PreviewSelectFieldWidget
 from plone.autoform import directives as form
-from plone.supermodel import model
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zope import schema
 
@@ -25,17 +24,7 @@ class GalleryTile(BaseImagesTile, DisplayTypeTileMixin):
     render = DisplayTypeTileMixin.render_display
 
 
-class IGalleryTileSchema(model.Schema):
-
-    form.widget('images', ImageRelatedItemsFieldWidget)
-    images = schema.List(
-        title=u"Images",
-        description=u"Select images or folders of images to display in "
-                    u"gallery",
-        value_type=schema.Choice(
-            vocabulary='plone.app.vocabularies.Catalog'
-        )
-    )
+class IGalleryTileSchema(IImagesTileSchema):
 
     form.widget('display_type', PreviewSelectFieldWidget,
                 tile_name='gallery')